Update Viewing_Tonight.py
added basic html improvments:
- added link to each hour of viewing
- added link from each hour back to top
- added each color coded by hour

it occurs to me that I should also make the background black and the text white as the intention is for this to be viewied at night or in some darkeness where as if I print it out, I would reverse that.  hmm, perhaps I need a printing out option.