Update Viewing_Tonight.py
note, none of the gui stuff is working fully as the location class uses the hard coded location.json.  I need to review the gui stuff to do a lot of work but first among that is to get it to match the location.json.  there is a bunch that I think i should detail in the readme or something